Change default for cutting_blade_usage_time
